Fixes fla and adds an InsanityException check for empty case names.
These changes were prompted by Fla., which was silently returning some
items with no case names. This caused confusing crashes in the the save
routine when it tried to save the empty values. In this commit I add
a check so that this should no longer happen, and I fix Fla so it
punts on items with no case name values.