Allows the caller to set the object that will do the case name tweaking, reducing the tests to about 20s instead of about 90.
The basic problem here was that to create short case names, we needed to load a collection of geographies from disk. This was done by a CaseNameTweaker object. That's fine and all, but it's a big list and doing it 200 times (once per scraper) takes a lot of time. The new system allows the caller to optionally provide a CaseNameTweaker instance to the Site object when it is initialized. This way, the caller can initialize a CaseNameTweaker object once and then pass it to all 200 scrapers, rather than having every scraper create a CaseNameTweaker object for itself.

This should have been an easy change, but unfortunately we didn't set up all of the subclasses of the AbstractSite to accept and then pass *args and **kwargs. If we had done that at the get go, this would have been easy, but instead I had to add this bit of code to every scraper using a regex (let's hope it worked properly!). In the future changes like this will be easier.

If you want to see the logic of the changes, look in AbstractSite and in tests.py. The rest is the same simple change over and over....