Makes illctapp inherit from ill.
I try hard to avoid repetition, so after I got the paths tweaked to my satisfaction and had illctapp working independently, I isolated the parts of it that were unique and brought them inline with those in ill. With that complete, I was able to make it inherit completely from ill, only customizing those parts that were unique. DRY!