Should make the case ID parameter work properly in GET requests.
This is a bit of a strange tweak, and I admit I don't particularly love it. From the research I've just done, it appears that case_id is the correct parameter to send with a GET request, and caseid is the correct parameter to send with a POST. This was completed using the following two URLs:

1. https://ecf.wyb.uscourts.gov/doc1/2061538313?case_id=32100&got_receipt=1
2. https://ecf.wyb.uscourts.gov/doc1/2061538313?caseid=32100&got_receipt=1

Number 1 works. Number 2 does not. However, if we remove the got_receipt parameter, we land at a page that has a button which sends a post request instead. Clicking the button on that page shows the following parameters:

    caseid: "32100"
    got_receipt: "1"

So, weirdly enough, PACER is inconsistent between POST and GET request parameters. Ugh. OTOH, so it goes.