Removes py3 compatibility. Neither requests-mock nor jsondate are compatible, and it just doesn't feel like our hill to die on.
I'm super frustrated with this, but my options are:

 - Drop support for py3,
 - Drop these two dependencies and figure out how to work things without them,
 - Rewrite these two dependencies.

We need a hero, but I'm not that hero, and thus won't be doing the rewrite. So, I can either rewrite my code, which works perfectly right now, or I can drop py3. Ugh. I'll pay the piper another day, and hopefully by then more libraries will have done the needed work to support py3. This shouldn't be an issue anymore, darn it.