Moves login and re-login code into the session. Contains breaking changes.
This is a fairly big and somewhat risky change to the session object so that it incorporates the login code which previously lived at the top level. The main benefit from this change is that we can easily do automatic re-login when our session expires. This is accomplished by checking the text returned by each query. If it contains the logout link, then we assume we're logged in. If not, we assume we need to log in.

The breakages caused by this are:

 - Login is no longer needed before creating a PacerSession object, since it can happen automatically.
 - PacerSessions now accept username and password fields (but do not accept pacer_token fields anymore).
 - The login and login_training (and a few other methods) are moved into the session object.

 It's a nice if tricky improvement.