Str fields now return the empty string instead of None in PACER scrapers.
This is an important change that complies with Django's philosophy on this topic. The creators of Django believe that it's better to avoid ambiguity around variables that do not have values. In the case of strings, they believe it's best to return the empty string — always — instead of sometimes returning a string and sometimes returning None. This change makes that happen at least for the Docket scraper.