feat(pacer): Adds a system for identifying invalid dockets before parsing
There are a variety of things that can be shown instead of dockets. This chunk of code ensures that
if we get those kinds of things, we don't parse them. Instead, we return an empty dict indicating
the item is invalid.

Another approach could be to raise an exception when this happens. That could work too, but I'm not
sure it makes a huge difference to the caller.