tests: Add some faked-up RSS dockets
These {are|will be} submitted by the @big_cases bot, as it goes
through RSS, which has limited information. As such, it submits only the
information it has, in a carefully crafted docket.

Consequently, in the test driver, don't require the date filed
information, because we just don't know it.

Two examples for two variations.

mad.html (Narvaez v. BerryHill): the case caption is included, because
the docket did not previously exist in CourtListener.

dcd.html (United States v. Manafort et al): The Robert Gates
subcase. The case caption is excluded, because (now) the case exists
in CourtListener.