doc fixes, rss tests, and a command-line driver (#203)
* README.rst: How to run stuff

How to run sample_caller w/o installing
How to run tests
How to run nosetests

* pacerdocket.py: Command-line PACER parser driver

  For running district court dockets.

* log_tools.py: Diagnostics to stderr, not stdout!

* README.rst: How to run packerdocket.py

* test_pacer: run_parsers_on_path() knob for requiredFields

Allow some variation in the required fields, because sometimes we may
not know the date filed, such as in faked-up RSS. This wounld't be a
good default, hence the knob.

* tests: Add some faked-up RSS dockets

These {are|will be} submitted by the @big_cases bot, as it goes
through RSS, which has limited information. As such, it submits only the
information it has, in a carefully crafted docket.

Consequently, in the test driver, don't require the date filed
information, because we just don't know it.

Two examples for two variations.

mad.html (Narvaez v. BerryHill): the case caption is included, because
the docket did not previously exist in CourtListener.

dcd.html (United States v. Manafort et al): The Robert Gates
subcase. The case caption is excluded, because (now) the case exists
in CourtListener.

* test_pacer: requiredFields -> required_fields

Too much time in Javascript.

* test_pacer: test_district_court_dockets -> test_fakerss_court_dockets

Whoops. Let's not have two test functions with the same name, please?

* README.rst: clarify `nosetests` non-dep

That is, we don't need nosetests, but tell people where to get it if
they want to use it.

* fake-rss: clarify recapupload.py's not ours

It's part of Big Cases.
So link to it in a comment.
(Or to the pending pull request)