fix(PACER): Fixes login so it works for filing accounts
I spent far too much time on this today. The crux of it is that the
standard login page uses gets really funky when you have a filing account.
Instead of simply doing a POST like a normal login page, it does a bunch
of junk involving sessions so that you can do a POST with your credentials,
then agree to a redaction notice via a separate page that pops up in a
modal, then get your cookies. Getting this all to work properly is
very hard, and I expect if I looked at it long enough, I'd discover that
their implementation was insecure. Sooo.....

Rather than continue down that road, I moved the login code over to cand's
login page, which operates in the same way as the training site. That meant
that we could ditch a fair bit of code, and just use the training site's
code. Great! But, this will fail when the last court goes to NextGen.