fix(pacer): Makes docket regex matching more effective
First, I loosened one of our docket number regexes so that it worked
better on the docket history page, but in doing that, I broke the docket
parser.

To fix that, I switched the order of the two docket number regexes on
the docket parser, so that the most specific one goes first, and the
less specific one (which is more likely to over-match) goes second.

Between these two changes, it looks like all our old tests pass and we
can more effectively capture docket numbers and case names on the
docket history report.