feat(PACER Hidden APIs): Better docket selection in PossibleCaseNumberApi
Adds new arguments that can be used to identify the correct docket:

 - criminal_or_civil: Which filters on whether there is a cv or a cr in
   the docket numbers returned.

 - office_number: Which filters based on the leading digit in a full
   docket number, e.g. the 1 in 1:23-cv-00002.

Also adds a new technique for identifying the correct case just by
pacer_case_id itself. If the above approaches have failed to find the
right item, we check if the remaining items have sequential
pacer_case_id's. If so, we pick the lowest one. If not, we proceed to
match based on case_name, as previous.