fix(PACER Bankr. Parsers): Better whitespace handling
Tweaks the parser to handle non-breaking spaces as normal spaces. This,
broke a few of the regexes, and uncovered some parsing errors on some of
the TXNB dockets.

Also includes changes to the blacklist of words that aren't allowed in
judge names due to some extra words that are now caught in the judge
field (previously, these were accidentally excluded by non-breaking
spaces).