fix(PACER Parsing): Adds support for service lists
Service lists have been causing us trouble for a while in the form of
crashing the code on CourtListener. This fix ensures that the type field
is populated when parsing service lists, which will make CL happy.

This is a bit funky because it makes the name of the party into
"Service List", but that essentially mirrors what's on PACER anyway. The
better fix might be to pull the service list into its own part of the
returned object, but that's a lot more work. And for what gain?

Closes #209