fix(PACER Parsing): Adds a block for the MetaMask conflicting extension
Blocking the extension after parsing would have been easy: I could have
just added the bad string to the list of error_strings and been done
with it. Alas, for this fix to work, we need to block even parsing or
else we have ten second parses for every chunk of HTML somebody with
this extension uploads.

So...this re-architects things so that our previous error string checker
is now upstream of the HTML parser. Docket error checking should be
unaffected, and anything that inherits from BaseReport should now have
one ERROR_STRING that they check for, which is the one matching
MetaMask.

Also includes tests, naturally.

Affects: freelawproject/recap#239.