feat(PACER Parsing): Support for dockets with "View Multiple Documents"
There's a checkbox in the bankruptcy courts that says:

 [ ] View multiple documents

If you check the box, you get a column of checkboxes on the docket that,
when checked, allow you to download a bunch of documents all at once.

That's fine, but our xpath was a bit too eager and included content from
the table that had the buttons to download this content. This commit
limits our xpath to exclude that table.

Too many corner cases!