fix(PACER Parsing): Handles count and offense info in the party listings
Some dockets list things like:

 - Pending counts
 - Highest offense level
 - Terminated counts
 - Highest offense level (Terminated)
 - Complaints

For these dockets, it's very hard to sort out what's what because the
formatting of this information is nearly identical to that of the
parties. The solution is to look for the word "Disposition" in the third
column and nuke based on that, which is effectively what this commit
does.