refactor(Case Name Shortener): Lazifies initiation of CaseNameTweaker
CaseNameTweaker is used to make very short names from case names. For
example, "United States v. Plymouth Rock" becomes just "Plymouth".

To do this, the Blue Book requires that you eliminate states, cities,
counties, etc, which this code does by reading those values from disk.
Doing that is slow, so we used to do that only during class initiation,
but that's annoying because sometimes you need the class, but you don't
use it in the code that's actually run. This change makes that
initiation of those placenames lazy such that it's only done the first
time the CaseNameTweaker object is *used*, not when it is initiated.