feat(IA Parsing): Adds a new parser for Internet Archive XML dockets
This is useful because it allows people to use Juriscraper for either
HTML from PACER or for XML from Internet Archive dockets. Regardless of
which is used, you'll get nearly the same friendly JSON object once
parsed, and you'll have the same interface while parsing it.

The only major difference between the items that will come back from the
IA parser and the PACER parser is the presence of attachment numbers
(and objects) in the IA XML parser.

For CourtListener, I expect this will result in a vastly simplified
code base, and I expect it'll be helpful for others as well that are
attempting to understand the IA docket XML.