fix(docket parsing): Improvements to case title identification
Now works better when the case title has an uppercase V. in it instead
of a lower case one.

This is implemented in a bit of a funny way, by trying to find the
lower case v, then the uppercase, then looking for in re cases. The
reason for this is that I don't want to have false positives for cases
containing strings like "V.R. Manufacturing", which might otherwise get
caught by the regexes.