fix(attachment page parsing): Fixes weird att page from mad
Looks like the first document in this entry is sealed. Normally such
document still show up OK as a row, but apparently mad takes the
approach of just leaving the row off the attachment page entirely. When
that happened, we got buggy because we were still looking for that row.