feat(appellate parsers): Adds the first four sample files
These are the Hawaii v. Trump travel ban case and the D'Apuzzo class
action. I chose these because one is big and important and the other is
of national importance. These are from two circuits, and I've got the
full and small version of each docket report so we have a lot to work
with.