feat(docket parser): Criminal data seems to be working properly.
This one is different than any I've made before (figures). The criminal
data is organized into HTML rows such that you'll have a tr for the
header info, a handful of trs for the actual information, and then the
next thing you'll have is a tr for the next section.

To combat this, you cannot use simple tools that allow you to look one
iteration ahead, and you cannot simply look at it one row at a time.
Instead, you have to keep track of what you've seen so far so that you
know what section you're currently working on.

I think I pulled off that design pretty well (it works and seems tidy
enough), but it's still kind of horrific to store so much state while
iterating a loop. Still, not sure there's any other way to do it.