fix(criminal parser): Adds more robustness.
Lots of issues resolved here, but main one was that we were only
counting parties properly if they had representation. This hit the wall
when we looked at the Michael Cohen case, which has a Special Master. It
also turns out that the last time we saw it, we messed it up too, but at
that stage in the litigation, the criminal data got associated with
Cohen. That *looked* right and so it slipped by, but in fact, the data
*should* be associated with the search warrants, as in the original
HTML.

All of this to say: This makes the criminal data associate with the
right party.