rss_feeds.py/docket_entries(): Document regexs, function
Give an example of what the regexs are matching and explain their
choices. In the abstract, it's really tough to debug and figure out
what's going on without this kind of thing.

Also, clarify the docstring of docket_entries() and the nonobvious
peculiarity that it does not return minute orders without attached
documents.