feat(rss feeds) Re-work the command line args
* rss_feeds.py: Allow reading from stdin

This is important to being able to debug the parser without repeatedly
pounding on the server, as well as being able to give it XML that the
server would never produce, exactly (e.g. a limited set with a single
specific entry, or maybe a specific type of entry)

* rss_feeds.py: Put _main() in its own function

* rss_feeds.py: Use argparse.

This python module is kind of a shitshow, but hey...

* rss_feeds.py: Allow -b commandline for bankruptcy

Because the parser is a little more flexible for bankruptcy cases.

* rss_feeds: prefer a <= foo <= b to  foo >= a and foo <= b

* Longer fake court_ids

Might as well make the district vs. bankruptcy part clear enough.
Hey, this'll break with appellate, too. Was it really thought through?