fix(pacer session): Updates session handling for newer versions of PACER
In old versions, PACER would set a cookie via JavaScript embedded in
the HTML using something like:

<SCRIPT LANGUAGE="JavaScript">
    document.cookie="PacerSession=xyz; path=/; domain=.uscourts.gov; secure;";
    if ("".length > 0) {
        document.cookie="; secure;";
    }
    if ("PacerPref=receipt=Y; path=/; domain=.uscourts.gov".length > 0) {
        document.cookie="PacerPref=receipt=Y; path=/; domain=.uscourts.gov; secure;";
    }
    if ("".length > 0) {
        document.cookie="; secure;";
    }
    if ("".length > 0) {
        document.cookie="; secure;";
    }
    if ("".length > 0) {
        location.assign("");
    }
</SCRIPT>

That was annoying because we had to parse it out in order to set the
session's cookie. Good news! Newer versions of PACER now set cookies
using HTTP headers (like a normal website), and so we can just login,
check if the cookie is there, and move on if so. If the cookie is
missing, we still fall back to the old method of parsing it from the
HTML.