fix(appellate parser): Ensure that we always have a document_number.
Some courts don't seem to use the document numbers. We need *something*,
so if we have a pacer_doc_id, but lack a document_number, use the
pacer_doc_id for the number.

This is...not ideal, but it's how RECAP has historically solved this
problem. This could also probably be solved by the caller rather than in
the parser, but this keeps the JSON data standardized from the parser,
which is worth something, even at the cost of some accuracy in the
parser results.