feat(pacer sessions): Removes support for training website
The training website, as far as I understand it, is the only
jurisdiction that no longer works after you have logged into the
NextGen log in. As such, it would require its own code paths that aren't
worth the effort to maintain and I've excised them in this commit. In my
next commit, I will be upgrading sessions to use the NextGen login,
since it's finally gotten to where things break if you don't log in that
way (specifically the hidden docket number look up API breaks).