feat(pacer sessions): Rewrites session to work with NextGen and filing accounts
This is rather annoying, but became necessary because the hidden API
for looking up docket numbers only worked if you had the correct
NextGenCSO cookie value, which you only got if you logged in via the
NextGen login thing. If you didn't do that, you'd instead get a HTML
with the XML embedded in it when you looked up docket numbers. That
HTML in turn had JavaScript embedded in it, presumably to set the
NextGenCSO cookie that you lacked. It's special.