fix(pacer query tests): Makes test pass without using training site.
I removed the training site login code in a previous commit. This
removes tests that relied on it and refactors them to use cand instead.

This also removes the VCR cassettes because they made me crazy and cost
me much more time than they ever saved.

The big downside here is that the full test suite now costs money. So be
it.