Version bump — it looks like micro versions aren't supported by pip?
Hard to say. The version appears on the pip website just fine:

https://pypi.org/project/juriscraper/1.19.9.1/

But then when I install/upgrade it doesn't get listed. It doesn't even
show up as a real version when supplying the install with the -vvv flag.

So I don't know what's up with this, but I guess pip has opinions and
one such opinion is that microversions aren't allowed. I guess.