fix(docket parser): Fixes parsing when bold-italics under party name
The bug here is that in some cases the "extra info" they put under
somebody's name on a non-bankruptcy docket can have bold italics and we
previously used that to identify the party type information on
bankruptcy dockets. So...make sure we only use that bankruptcy-related
heuristic on bankruptcy dockets, not regular ones.