fix(pacer session): Ensure we log in when hidden API calls require it
I'm guessing this varies across versions of PACER, but we've had an
issue that our session wasn't getting automatically re-logged in when
using the PossibleCaseNumberApi. The way we were sniffing for this
before was over-inclusive because it checked for the presence of several
different text strings and assumed if they were there, we made a
successful request.

Well, it turns out that those are there even when you're logged out, so
we weren't logging back in like we should have been. This was certainly
tested before, so I suspect this is related to a change in PACER itself.

In any case, the way this works now is that we run all the checks we had
previously, but then have one more check to see if we're getting a
message that we're logged out. If we see that message, we make sure to
login again.