fix(docket parser): Support judges with "v." as middle initial
Fun one. What was happening here is that our regex would match on the
judge's name when trying to find the case name, with the result that the
parser would have two things it thought were the case name. When that
happened, it'd give up since it didn't know which one was right.

The code here was originally written to be super aggressive about
finding matches anywhere in the docket metadata. Turns out that now that
we have lots of good tests, we can safely stop doing that, and be
certain that the case name is in the first element.