fix(pacer session): Clear cookies before trying to log in again
This will ensure that the login works properly. Without clearing them,
you run the risk of failing to get the login page because you're not
quite logged out. Most recently, this happened when we got a page that
said, "This page will expire in...[so many minutes]." When that
happened, we tried to login again, were still logged in, and then had
issues.

I think this will fix that, but I won't know for sure until we have an
old session that needs to be refreshed (like this one did). That'll
happen tonight if this code is deployed.