feat(pacer): Factor out and generalize the label and tail code
@johnhawkinson has some good code that looped over label headers and
assigned their tails to to the header in a nice dict. This takes that
code from his work, factors it out, and lifts it to the base class so
that the code can be used in other places. (In particular, I am using
this in the upcoming bankruptcy claims register report.)