fix(lasc): Make parser more robust for different data sourcing
I'm not sure why this is different, but we have a few million local
docket JSON files on disk that we want to parse, and for some reason
they don't have quite the same format as those we get back from the
site itself. Luckily, handling this is pretty easy; we just need to get
the right data at the outset and the rest falls into place.