fix(lasc): Clean up all lasc data points using standard PACER tool
This is an aggressive cleaning tool that does a number of cleanup tasks.
I debated using this at all, but after seeing some of the results coming
in, it seemed like some clean up would be a net benefit. Even looking at
the three sample files from our tests, this is borne out with a 100%
improvement in their white space.