fix(pacer): Switch parameter name from case_id to caseid
It took me a long time to realize it today, but when this is wrong,
it's impossible to get PDFs with their headers on them. I was surprised
to see this field be wrong because all the other fields on this POST
request use snake case, but I checked at DCD, CAND, and ILND and they
all used caseid instead of case_id.

It's quite possible that PACER is inconsistent at times in how it cases
variable names.