feat(pacer): Include time and tz in date_filed fields
This is important so that we can identify when two entries come at
the exact same moment, as can happen in multi-event entries.
Previously, such entries would have the same date, but that's not as
good as having the same date and time.

Note that we do one strange thing here. feedparser dates have parsed
fields that should provide decent values. For example, prior to this
change, we used the published_parsed field. Well, that field returns
a time struct, which basically normalizes the timezone information
and converts it into a naive datetime. That's lame and weird, and
I've filed a bug about it here:

    "Why is published_parsed naive?"
    https://github.com/kurtmckee/feedparser/issues/212

The maintainer has responded that yes, this is weird and it's because
feedparser is 20-something years old.

So, instead of using that weird value, we parse the damned thing
ourselves using dateutil, which gives us a tz-aware datetime object.
Better!