fix(iquery): Fix unicode error
From what I can tell, when you run text_content() on an lxml node, it generally
prefers to give you a str. Except when there's unicode content in there, in
which case it gives you a unicode object. This fix makes sure that whenever it
gives you, we convert it into unicode.

What prompted this was the following code in the HTML:

<BR><B>Debtor&nbsp;discharged:</B>&nbsp;11/01/2018

That dumb &nbsp; got converted into `\xa0` in unicode, which barfed the system
with:

Traceback (most recent call last):
  File "/usr/local/lib/python2.7/site-packages/celery/app/trace.py", line 385, in trace_task
    R = retval = fun(*args, **kwargs)
  File "/usr/local/lib/python2.7/site-packages/celery/app/trace.py", line 648, in __protected_call__
    return self.run(*args, **kwargs)
  File "/opt/celery/cl/scrapers/tasks.py", line 638, in update_docket_info_iquery
    d = update_docket_metadata(d, report.metadata)
  File "/usr/local/lib/python2.7/site-packages/juriscraper/pacer/case_query.py", line 166, in metadata
    self._get_label_value_pair(bold, True, field_names)
  File "/usr/local/lib/python2.7/site-packages/juriscraper/pacer/docket_report.py", line 233, in _get_label_value_pair
    label = label.decode("utf-8")
  File "/usr/local/lib/python2.7/encodings/utf_8.py", line 16, in decode
    return codecs.utf_8_decode(input, errors, True)
UnicodeEncodeError: 'ascii' codec can't encode character u'\xa0' in position 6: ordinal not in range(128)

Hopefully this fixes that once and for all.