fix(pacer): Better judge handling on dockets
Moves regex handling out of xpath (!) and into Python, thus fixing two
issues. First, we make sure we split on newlines where we should, so that
we don't over-include, adding "Referring" to the end of Assigned to judge
names.

Second, we ignore newlines in the middle of judge names, so they don't
create problems.

Thank goodness for the test suite.