feat(pacer): Land improved cleaner
This fixes two issues:

 - Performance issues caused by random stuff (CSS?) that was occasionally in
   our HTML
 - Bad script removal being caused by the old method.

Unfortunately, this fix took several weeks to land because LXML had a big bug
in the Cleaner() class that had either be fixed or worked around. As I worked
on that, I admit I sort of lost track of why I was working on this. I know
that there is an example file somewhere that made the parsers very slow and
that the solution was to use the Cleaner, but I can't find that file anymore.

I'm a bit frustrated, but so it goes. The upside is that along the way, this
fixes the parser to make it more accurate (see tests in next commit).