fix(pacer): Make inheritance explicit
Previously, this relied on the very-handy super infrastructure, but
that led to confusion because it wasn't always clear which function was
super to another when doing multiple inheritance. For example:

  class Baz(object):

  class Bar(object):

  class Foo(Bar, Baz):

Assuming Foo, Bar, and Baz all have __init__ methods with super calls,
initiating Foo as above would: that class would:
 - run the __init__ method of Foo
 - run the __init__ method of Bar
 - run the __init__ method of Baz

That's OK, except that if you want to *sometimes* provide arguments to
the init methods, it's going to get ugly because in the above case,
super of Bar is Baz (due to Foo's multiple intheritance). But in this
case, super of Bar is object (due to single inheritance):

  class Bazinga(Bar):

So, if super of Bar is sometimes Baz and sometimes Bazinga, it gets
difficult to provide the right number of parameters to Bar's super call.

This fixes all that by being explicit and dropping super.