fix(nyappterm): Better cleanup for HTML
The previous version would return HTML like this:

    <partyblock>
    ...stuff...
    </partyblock>

Which was kind of OK (browsers ignore unknown elements), but when we tried to
sniff the extension, that made our sniffer think it was a WPD file, which was
not so good.

The new code just returns the "...stuff..." part, and our sniffer properly
handles that.

Fixes: COURTLISTENER-2T