feat(pacer): Better case name parsing
This implements a fall back for when the case name can't be pulled from the
case metadata. The fall back is pretty naive, but it'll catch some instances
that we currently miss. The basic idea is that if there are clearly labeled
plaintiffs and defendants, we might as well use the first plaintiff and
defendant to make a case name.