fix(reqs): Update date util
Now that we're using poetry, it's becoming clear how broken our dependencies are. For example, when I just tried to install all the CL dependencies, I got this error:

```
Resolving dependencies... (87.0s)

SolverProblemError

Because juriscraper (1.28.16 git rev py38) depends on python-dateutil (2.5.0)
 and cl depends on python-dateutil (^2.8.1), juriscraper is forbidden.
So, because cl depends on juriscraper (1.28.16 git rev py38), version solving failed.
```

This means that prior to now, we've been ambiguous about which version of date util we use, and we just used whichever was installed first (which I think would be the lower version, because deps are installed alphabetically, and dateutil from CL would come before Juriscraper).

Anyway, this should fix that.