fix(warnings): Don't enable warnings as broadly
I'm not entirely sure what changes this will bring, but it does fix one
issue I've been wrestling with. Over in freelawproject/courtlistener#1441
I just did a binary search through all of CourtListener to figure out why
we were getting the following error all the time after switching to py3:

sys:1: ResourceWarning: unclosed file <_io.TextIOWrapper name='/dev/null' mode='w' encoding='UTF-8'>

In the end, it was this line in Juriscraper! I'm removing it. If this
means that logging needs further refinement here, we should do that, but
for now I need a fix to the other issue.