feat(pacer): Make download_pdf method return msg
This enhances PACER's download_pdf function to return an error message in
addition to either a None or a Response object. In so doing, it allows callers
to report the errors they run into. This is important in our PACER fetch API,
where users deserve to know why something failed.

Fixes: #510