fix(texag): Fix Texas AG decisions
It was retired in 2018 but appears to be getable again.

A backscraper may be warranted but they
publish a sidepanel with most recent opinions
so i think that we should just collect them
there.  As opposed to going thru the much
more cumbersome process.  Although it
also could be easy.