test_PacerParseACMSDocketTest.py: Add datatypes/repr() test
Add a test to look at the actual python repr() of the parsed data, not just the JSON conversion.

This is worth doing since the JSON conversion loses the datetime data
type information, about which we had a bug during
development:(11c425681be8d904e78d10aeaf325e703c44abf4, which is
probably an orphased SHA1 by the time this is merged, but is relative to
line 48 of acms_docket.py).