refactor(neb): Update Neb/NebCtApp
Python 3.11 and 3.12 parsed urls differently in lxml which was
causing a problem with one weird link in the nebraska pages
unrelated to opinions.

I overrode the return response text method to remove the links
before processing the html

Additionally, I refactored teh code to fix a few things
Identify status properly and fix parsing which was off
in some cases.