fix(ny, nyappdiv, nyappterm, nytrial): add X-Api-Token header
Fixes #1042

Add a function to set the required api token in both headers objects used in Juriscraper / Courtlistener. For this to work, we will need to modify CL to use the headers for ny courts.

Use that function on ny, nyappdiv, nyappterm, nytrial, which are base classes for more scrapers